requirements.txt
asyncpg>=0.29.0
anthropic>=0.34.0
pydantic>=2.0.0
python-dateutil>=2.8.0
fastapi>=0.111.0
uvicorn>=0.29.0
python-multipart>=0.0.9
httpx>=0.27.0
twilio>=8.0.0
boto3>=1.34.0
pyngrok>=7.0.0
